Refactor combat value calculations to use double for precision; update CombatUnit properties to double; enhance Battle class for improved statistics tracking; add utility method for console output of unit stats.

+++ OgameGenSim/Loader.cs
@@ -117,7 +117,7 @@ public class Loader(HttpClient client)
         return cleanUnits;
     }
 
-    private static float CalculateCombatValueWithBonuses(float defaultValue, double LFBonus, int techId, int techLevel, int charatcterClassId, int allianceClassId)
+    private static float CalculateCombatValueWithBonuses(double defaultValue, double LFBonus, int techId, int techLevel, int charatcterClassId, int allianceClassId)
     {
         if(charatcterClassId == (int)PlayerClass.General)
         {
